# golang-backend/docs/swagger.yaml
basePath: /
definitions:
  controller.BaseSwaggerResp:
    properties:
      code:
        example: 0
        type: integer
      msg:
        example: 操作成功
        type: string
      ts:
        example: 1700000000000
        type: integer
    type: object
...
  controller.SwaggerTunnelIDReq:
    properties:
      tunnelId:
        example: 1
        type: integer
    type: object
  controller.SwaggerTunnelIfaceReq:
    properties:
      iface:
        example: eth0
        type: string
      tunnelId:
        example: 1
        type: integer
    type: object
  controller.SwaggerTunnelPathSetReq:
    properties:
      path:
        example:
        - 2
        - 3
        items:
          type: integer
        type: array
      tunnelId:
        example: 1
        type: integer
    type: object
...
info:
  contact: {}
  description: 面板后端接口文档
  title: network-panel API
  version: "1.0"
paths:
...
  /api/v1/tunnel/cleanup-temp:
    post:
      consumes:
      - application/json
      parameters:
      - description: 隧道ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerTunnelIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 清理路径临时服务（iperf3）
      tags:
      - tunnel
...
  /api/v1/tunnel/iface/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: 隧道ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerTunnelIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 获取隧道节点出口IP
      tags:
      - tunnel
  /api/v1/tunnel/iface/set:
    post:
      consumes:
      - application/json
      parameters:
      - description: 隧道ID与出口IP
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerTunnelIfaceReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 设置隧道节点出口IP
      tags:
      - tunnel
...
  /api/v1/tunnel/path/check:
    post:
      consumes:
      - application/json
      parameters:
      - description: 隧道ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerTunnelIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 检查多级路径节点状态
      tags:
      - tunnel
  /api/v1/tunnel/path/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: 隧道ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerTunnelIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 获取隧道多级路径
      tags:
      - tunnel
  /api/v1/tunnel/path/set:
    post:
      consumes:
      - application/json
      parameters:
      - description: 隧道ID与路径节点
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerTunnelPathSetReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 设置隧道多级路径
      tags:
      - tunnel
...
swagger: "2.0"